COR: json_get_double prend en compte si la source est une string
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@5698 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/ClientGTK3/Include/trame.h
@@ -157,7 +157,7 @@
     JsonNode *cadran;
     gint   layer;                                                                      /* Groupe de deplacement du motif */
     gint   selection;
-    gfloat valeur;
+    gdouble valeur;
   };
 
  struct TRAME_ITEM_CAMERA_SUP
